requirements.txt
python-dotenv
beautifulsoup4
langchain
langchainhub
langchain-openai
langchain-community
langgraph
tavily-python
pytest
chromadb
tiktoken
langchain-chroma
langchain-tavily